% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthRateMethods.R
\name{birthDeathMCMC}
\alias{birthDeathMCMC}
\title{Growth rate estimate using MCMC}
\usage{
birthDeathMCMC(
  tree,
  maxGrowthRate = 4,
  alpha = 0.05,
  showProgress = TRUE,
  nChain = 3,
  nCores = 1,
  chainLength = 2000
)
}
\arguments{
\item{tree}{An ultrametric tree subset to include only the clone of
interest. Alternatively, a list with several such trees.}

\item{maxGrowthRate}{Sets upper bound on birth rate. Default is 4 but this
will depend on the nature of the data}

\item{alpha}{Used for calculation of confidence intervals. 1-alpha confidence intervals used with default of alpha = 0.05 (95 percent confidence intervals)}

\item{showProgress}{TRUE or FALSE, should the Rstan MCMC progress be printed?}

\item{nChain}{Number of chains to run in MCMC. Default is 3}

\item{nCores}{Number of cores to perform MCMC. Default is 1, but chains can
be run in parallel}

\item{chainLength}{Number of iterations for each chain in MCMC. Default is
2000, increase if stan tells you to}
}
\value{
A dataframe including the net growth rate estimate, confidence
intervals, and other important details (clone age estimate, runtime, n,
etc.)
}
\description{
Uses Rstan and the No U-turn sampler to approximate the likelihood
from Stadler 2009
}
\examples{
df <- birthDeathMCMC(cloneRate::exampleUltraTrees[[1]])

}
\seealso{
\link{internalLengths} \link{maxLikelihood} which use
alternative methods for growth rate estimation from an ultrametric tree.
}
