% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneRate-package.R
\docType{package}
\name{cloneRate-package}
\alias{cloneRate-package}
\alias{cloneRate}
\title{The 'cloneRate' package.}
\description{
Quickly estimate the net growth rate of a population or clone whose growth can be approximated by a birth-death branching process.
}
\references{
Johnson et al. 2023 (under review at Bioinformatics). bioRxiv preprint \url{doi:10.1101/2023.02.24.529817}
}
