% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTrees.R
\name{coal_to_tree}
\alias{coal_to_tree}
\title{Generate tree from coalescence times}
\usage{
coal_to_tree(coal_times)
}
\arguments{
\item{coal_times}{A numeric vector of coalescence times}
}
\value{
An ape object of class "phylo" representing the ultrametric
phylogenetic tree with edge lengths in units of time.
}
\description{
generates a tree from a vector of coalescence times by randomly
merging lineages.
}
\examples{
# Generate an ape phylo tree with n tips from a vector of n-1 coalescence times
randomCoalTimes <- c(9.3, 7.8, 10.15, 11.23, 9.4, 8.8, 10.01, 13)
tree <- coal_to_tree(randomCoalTimes)

}
