% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realData.R
\docType{data}
\name{embryonic_time_trees}
\alias{embryonic_time_trees}
\title{Embryonic time-based trees from human blood}
\format{
A \code{list} of containing trees with mutation-based edge lengths.
Two of these trees, CB001 and CB002, are from cord blood samples taken at birth.
The remaining trees are from adult samples, truncated to include only the first 55 mutations,
and the scaled to convert mutations to years
based on findings from Mitchell et al. in the cord blood samples.
Each element of the \code{list} is a tree as a class \code{phylo} object.
See ape package documentation for details on class \code{phylo} objects.
Names of each \code{phylo} object (tree) in the list matches the naming used
in the sources. Source paper is indicated in the metadata dataframe for each tree.
}
\usage{
data(embryonic_time_trees)
}
\description{
12 trees from 12 individual donors.
}
\examples{
# Plot embryonic time-based trees from cord blood sample CB002
ape::plot.phylo(cloneRate::embryonic_time_trees[["CB002"]],
  direction = "downwards", show.tip.label = FALSE
)

# Calculate the growth rate using the max likelihood method.
out <- maxLikelihood(cloneRate::embryonic_time_trees)

}
\references{
These datasets were generated and annotated in:
\href{https://pubmed.ncbi.nlm.nih.gov/35058638/}{Williams et al. 2022}
\href{https://pubmed.ncbi.nlm.nih.gov/35650442/}{Mitchell et al. 2022}
\href{https://pubmed.ncbi.nlm.nih.gov/35650444/}{Fabre et al. 2022}
}
\keyword{datasets}
