% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthRateMethods.R
\name{inputCheck}
\alias{inputCheck}
\title{inputCheck}
\usage{
inputCheck(subtree, alpha)
}
\arguments{
\item{subtree}{An ape tree subset to include only the clone of interest}

\item{alpha}{Used for calculation of confidence intervals. 1-alpha confidence
intervals used with default of alpha = 0.05 (95% confidence intervals)}
}
\description{
Check the validity of inputs to growth rate fns
}
