% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalLengths.R
\name{internalLengths}
\alias{internalLengths}
\title{internalLengths}
\usage{
internalLengths(subtree, withZeroes = F)
}
\arguments{
\item{subtree}{An ape tree subset to include only the clone of interest}

\item{withZeroes}{Boolean indicating whether the subtree includes a zeroes
outgroup as a tip}
}
\value{
A named vector including the net growth rate estimate and the sum of
internal lengths
}
\description{
Provides an estimate for the net growth rate of the clone with
    confidence bounds, as well as the sum of internal lengths
}
\examples{
data(exampleTrees)
output_vec <- internalLengths(exampleTrees[[1]])
}
