% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteFrequency.R
\name{siteFrequency}
\alias{siteFrequency}
\title{Get site frequency of a tree}
\usage{
siteFrequency(tree, includeStem = F)
}
\arguments{
\item{tree}{An ultrametric tree subset to include only the clone of
interest. Alternatively, a list with several such trees.}

\item{includeStem}{Boolean indicating whether we should count the stem of the tree as contributing to the site frequency distribution. Default is FALSE}
}
\value{
A data.frame with three columns: the number of descendant cells, site
frequency in units of time or mutations, and normalized site frequency. If a
list of trees are input, output will be a data.frame with two columns for
each input tree and a final column with averaged site frequency spectrum.
}
\description{
\code{siteFrequency()} returns a data.frame with the frequency of in units of time or mutations
}
\examples{
siteFrequency(exampleUltraTrees[[1]])

}
\seealso{
\code{\link[=internalLengths]{internalLengths()}}, \code{\link[=sharedMuts]{sharedMuts()}}
}
