% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneRate-package.R
\docType{package}
\name{cloneRate-package}
\alias{cloneRate-package}
\alias{cloneRate}
\title{The 'cloneRate' package.}
\description{
Quickly estimate the net growth rate of a population or clone whose growth can be approximated by a birth-death branching process.
}
\references{
Johnson et al. 2023 Bioinformatics. \url{doi:10.1093/bioinformatics/btad561}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bdj34/cloneRate}
  \item \url{https://bdj34.github.io/cloneRate/}
  \item Report bugs at \url{https://github.com/bdj34/cloneRate/issues}
}

}
\author{
\strong{Maintainer}: Brian Johnson \email{brian.d.johnson97@gmail.com} (\href{https://orcid.org/0000-0002-9477-4677}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Yubo Shuai [copyright holder]
  \item Jason Schweinsberg (\href{https://orcid.org/0000-0002-4152-0059}{ORCID}) [copyright holder]
  \item Kit Curtius (\href{https://orcid.org/0000-0002-2678-0960}{ORCID}) [copyright holder]
}

}
