% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalLengths.R
\name{internalLengths}
\alias{internalLengths}
\title{internalLengths}
\usage{
internalLengths(subtree, includeStem = F, alpha = 0.05)
}
\arguments{
\item{subtree}{An ape tree subset to include only the clone of interest}

\item{includeStem}{Boolean indicating whether we should count the stem of the
tree as contributing to the internal lengths summation}

\item{alpha}{Used for calculation of confidence intervals. 1-alpha confidence
intervals used with default of alpha = 0.05 (95% confidence intervals)}
}
\value{
A named vector including the net growth rate estimate and the sum of
internal lengths
}
\description{
Provides an estimate for the net growth rate of the clone with
    confidence bounds, as well as the sum of internal lengths
}
\examples{
data(exampleTrees)
output_vec <- internalLengths(exampleTrees[[1]])
}
