% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthRateMethods.R
\name{moments}
\alias{moments}
\title{moments}
\usage{
moments(subtree, alpha = 0.05)
}
\arguments{
\item{subtree}{An ape tree subset to include only the clone of interest}

\item{alpha}{Used for calculation of confidence intervals. 1-alpha confidence
intervals used with default of alpha = 0.05 (95% confidence intervals)}
}
\value{
A dataframe including the net growth rate estimate, confidence intervals,
and other important details (runtime, n, etc.)
}
\description{
Provides an estimate for the net growth rate of the clone with
    confidence bounds using the method of moments.
}
\examples{
data(exampleTrees)
df <- moments(exampleTrees[[1]])
}
