% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTree.R
\name{simTree}
\alias{simTree}
\title{simTree}
\usage{
simTree(a, b, cloneAge, n, precision = 10000)
}
\arguments{
\item{a}{Birth rate}

\item{b}{Death rate}

\item{cloneAge}{Clone age}

\item{n}{Number of samples (number of tips of the tree to be returned)}

\item{precision}{Rmpfr param for handling high precision numbers}
}
\value{
An ape tree object
}
\description{
Generates a sampled tree from a supercritial (birth rate >
    death rate) birth and death branching process according to the coalescent
    point process described in "Lambert, A. The coalescent of a sample from a
    binary branching process. (2018)."
}
\examples{
tree <- simTree(a=1, b=0.5, cloneAge = 20, n = 50)
}
